requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
pymongo==4.5.0
python-dotenv==1.0.0
python-slugify==8.0.1
google-generativeai==0.3.0
PyJWT==2.8.0
Werkzeug==2.3.7
bson==0.5.10
